Moved common window functionality to helper classes
