+++ src/Sidekick.Wpf/CloudflareWindow.xaml.cs
@@ -1,12 +1,11 @@
 using System.ComponentModel;
 using System.Windows;
-using System.Windows.Forms;
-using System.Windows.Interop;
 using System.Windows.Media;
 using Microsoft.Extensions.Logging;
 using Microsoft.Web.WebView2.Core;
 using Sidekick.Apis.Poe.Clients;
 using Sidekick.Apis.Poe.CloudFlare;
+using Sidekick.Wpf.Helpers;
 using Application=System.Windows.Application;
 
 namespace Sidekick.Wpf;
@@ -50,42 +49,9 @@ public partial class CloudflareWindow
             Background = (Brush?)new BrushConverter().ConvertFrom("#000000");
             Opacity = 0.01;
 
-            CenterOnScreen();
+            CenterHelper.Center(this);
             Activate();
         });
-    }
-
-    private void CenterOnScreen()
-    {
-        // Get the window's handle
-        var windowHandle = new WindowInteropHelper(this).Handle;
-
-        // Get the screen containing the window
-        var currentScreen = Screen.FromHandle(windowHandle);
-
-        // Get the working area of the screen (excluding taskbar, DPI-aware)
-        var workingArea = currentScreen.WorkingArea;
-
-        // Get the DPI scaling factor for the monitor
-        var dpi = VisualTreeHelper.GetDpi(this);
-
-        // Convert physical pixels (from working area) to WPF device-independent units (DIPs)
-        var workingAreaWidthInDips = workingArea.Width / (dpi.PixelsPerInchX / 96.0);
-        var workingAreaHeightInDips = workingArea.Height / (dpi.PixelsPerInchY / 96.0);
-        var workingAreaLeftInDips = workingArea.Left / (dpi.PixelsPerInchX / 96.0);
-        var workingAreaTopInDips = workingArea.Top / (dpi.PixelsPerInchY / 96.0);
-
-        // Get the actual size of the window in DIPs
-        var actualWidth = Width;
-        var actualHeight = Height;
-
-        // Calculate centered position within the working area
-        var left = workingAreaLeftInDips + (workingAreaWidthInDips - actualWidth) / 2;
-        var top = workingAreaTopInDips + (workingAreaHeightInDips - actualHeight) / 2;
-
-        // Set the window's position
-        Left = left;
-        Top = top;
     }
 
     protected override void OnClosing(CancelEventArgs e)
