+++ src/Sidekick.Wpf/Helpers/MaximizeHelper.cs
@@ -1,0 +1,131 @@
+﻿using System.Runtime.InteropServices;
+using System.Windows;
+using System.Windows.Interop;
+
+namespace Sidekick.Wpf.Helpers;
+
+// ReSharper disable All
+
+/// <summary>
+/// Code to make maximizing the window take the taskbar into account. https: //stackoverflow.com/questions/20941443/properly-maximizing-wpf-window-with-windowstyle-none
+/// </summary>
+public static class MaximizeHelper
+{
+    public static void AddHook(MainWindow window)
+    {
+        var hwndSource = PresentationSource.FromVisual(window) as HwndSource;
+        if (hwndSource == null)
+        {
+            return;
+        }
+
+        hwndSource.AddHook(HookProc);
+    }
+
+    public static void RemoveHook(MainWindow window)
+    {
+        var hwndSource = PresentationSource.FromVisual(window) as HwndSource;
+        if (hwndSource == null)
+        {
+            return;
+        }
+
+        hwndSource.RemoveHook(HookProc);
+    }
+
+    private static IntPtr HookProc(IntPtr hwnd, int msg, IntPtr wParam, IntPtr lParam, ref bool handled)
+    {
+        if (msg != WM_GETMINMAXINFO)
+        {
+            return IntPtr.Zero;
+        }
+
+        // We need to tell the system what our size should be when maximized. Otherwise it will
+        // cover the whole screen, including the task bar.
+        var mmi = (MINMAXINFO)Marshal.PtrToStructure(lParam, typeof(MINMAXINFO))!;
+
+        // Adjust the maximized size and position to fit the work area of the correct monitor
+        var monitor = MonitorFromWindow(hwnd, MONITOR_DEFAULTTONEAREST);
+
+        if (monitor != IntPtr.Zero)
+        {
+            var monitorInfo = new MONITORINFO
+            {
+                cbSize = Marshal.SizeOf(typeof(MONITORINFO)),
+            };
+            GetMonitorInfo(monitor, ref monitorInfo);
+            var rcWorkArea = monitorInfo.rcWork;
+            var rcMonitorArea = monitorInfo.rcMonitor;
+            mmi.ptMaxPosition.X = Math.Abs(rcWorkArea.Left - rcMonitorArea.Left);
+            mmi.ptMaxPosition.Y = Math.Abs(rcWorkArea.Top - rcMonitorArea.Top);
+            mmi.ptMaxSize.X = Math.Abs(rcWorkArea.Right - rcWorkArea.Left);
+            mmi.ptMaxSize.Y = Math.Abs(rcWorkArea.Bottom - rcWorkArea.Top);
+        }
+
+        Marshal.StructureToPtr(mmi, lParam, true);
+
+        return IntPtr.Zero;
+    }
+
+    private const int WM_GETMINMAXINFO = 0x0024;
+
+    private const uint MONITOR_DEFAULTTONEAREST = 0x00000002;
+
+    [DllImport("user32.dll")]
+    private static extern IntPtr MonitorFromWindow(IntPtr handle, uint flags);
+
+    [DllImport("user32.dll")]
+    private static extern bool GetMonitorInfo(IntPtr hMonitor, ref MONITORINFO lpmi);
+
+    [Serializable]
+    [StructLayout(LayoutKind.Sequential)]
+    public struct RECT
+    {
+        public int Left;
+        public int Top;
+        public int Right;
+        public int Bottom;
+
+        public RECT(int left, int top, int right, int bottom)
+        {
+            this.Left = left;
+            this.Top = top;
+            this.Right = right;
+            this.Bottom = bottom;
+        }
+    }
+
+    [StructLayout(LayoutKind.Sequential)]
+    public struct MONITORINFO
+    {
+        public int cbSize;
+        public RECT rcMonitor;
+        public RECT rcWork;
+        public uint dwFlags;
+    }
+
+    [Serializable]
+    [StructLayout(LayoutKind.Sequential)]
+    public struct POINT
+    {
+        public int X;
+        public int Y;
+
+        public POINT(int x, int y)
+        {
+            this.X = x;
+            this.Y = y;
+        }
+    }
+
+    [StructLayout(LayoutKind.Sequential)]
+    public struct MINMAXINFO
+    {
+        public POINT ptReserved;
+        public POINT ptMaxSize;
+        public POINT ptMaxPosition;
+        public POINT ptMinTrackSize;
+        public POINT ptMaxTrackSize;
+    }
+
+}
